LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-fallback-scroll-expected.txt

XXX

FAIL @container anchored() without applied fallback at initial scroll position assert_equals: expected "no" but got ""
FAIL @container anchored() applies fallback after scrolling down assert_equals: expected 200 but got 0
